Transformando o arquivo server.js para um .ts e refeito as rotas de board
